% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_col_name_function.R
\name{clean_col_name}
\alias{clean_col_name}
\title{A Cat Function}
\usage{
clean_col_name(col, lowercase = TRUE, uppercase_id = TRUE, prefix = FALSE)
}
\arguments{
\item{col}{The column name you're cleaning}

\item{lowercase}{Convert the column name to lower case. Default: True}

\item{uppercase_id}{Make any instance of "id" uppercase. Default: True}

\item{prefix}{Prepend the string in prefix. Defaults FALSE}
}
\description{
This function cleans column names
}
\examples{
clean_col_name()
}
\keyword{cleaning}
\keyword{column}
\keyword{names}
