% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_col_name_function.R
\name{clean_col_name}
\alias{clean_col_name}
\title{A Cat Function}
\usage{
clean_col_name(col)
}
\arguments{
\item{col}{The column name you're cleaning}
}
\description{
This function allows you to express your love of cats.
}
\examples{
clean_col_name()
}
\keyword{cats}
